+++ datadog-checks-base/datadog_checks/checks/prometheus/mixins.py
@@ -4,6 +4,8 @@
 
 from fnmatch import fnmatchcase
 import requests
+from urllib3 import disable_warnings
+from requests.packages.urllib3.exceptions import InsecureRequestWarning
 from collections import defaultdict
 from google.protobuf.internal.decoder import _DecodeVarint32  # pylint: disable=E0611,E0401
 from ...utils.prometheus import metrics_pb2
@@ -463,6 +465,9 @@ class PrometheusScraper(object):
         verify = True
         if isinstance(self.ssl_ca_cert, basestring):
             verify = self.ssl_ca_cert
+        elif self.ssl_ca_cert is False:
+            disable_warnings(InsecureRequestWarning)
+            verify = False
         try:
             response = requests.get(endpoint, headers=headers, stream=True, timeout=1, cert=cert, verify=verify)
         except (requests.exceptions.SSLError):