+++ tasks.py
@@ -1,7 +1,9 @@
 from __future__ import print_function, unicode_literals
 
+import fnmatch
 import json
 import os
+import shutil
 import uuid
 from collections import OrderedDict
 from io import open
@@ -15,12 +17,15 @@ ROOT = os.path.dirname(os.path.abspath(__file__))
 AGENT_BASED_INTEGRATIONS = [
     'datadog-checks-base',
     'disk',
+    'directory',
     'envoy',
     'istio',
     'kube_proxy',
     'kubelet',
     'linkerd',
+    'nfsstat',
     'prometheus',
+    'redisdb',
     'vsphere',
 ]
 
@@ -148,6 +153,40 @@ def upgrade(ctx, package=None, version=None, verbose=False):
                     'requirements.in',
                     hide='both'
                 )
+
+                
+@task(help={})
+def cleanup(ctx):
+    """
+    Cleanup all temporary files and directories
+
+    Example invocation:
+        inv cleanup
+    """
+
+    should_delete_directories = [
+        ".eggs",
+        ".pytest_cache",
+        ".tox",
+        "__pycache__"
+    ]
+    should_delete_directories_globs = [
+        "*.egg-info"
+    ]
+
+    for path, dirnames, filenames in os.walk('.'):
+        for filename in fnmatch.filter(filenames, '*.pyc'):
+            file_path = os.path.join(HERE, path, filename)
+            os.remove(file_path)
+        for dirname in dirnames:
+            if dirname in should_delete_directories:
+                dir_path = os.path.join(HERE, path, dirname)
+                shutil.rmtree(dir_path)
+
+        for dir_glob in should_delete_directories_globs:
+            for dirname in fnmatch.filter(dirnames, dir_glob):
+                dir_path = os.path.join(HERE, path, dirname)
+                shutil.rmtree(dir_path)
 
 
 @task(help={