+++ vsphere/datadog_checks/vsphere/vsphere.py
@@ -205,11 +205,13 @@ class VSphereCheck(AgentCheck):
 
     def _get_server_instance(self, instance):
         i_key = self._instance_key(instance)
+        tags = instance.get('tags', [])
 
         service_check_tags = [
             'vcenter_server:{0}'.format(instance.get('name')),
             'vcenter_host:{0}'.format(instance.get('host')),
-        ]
+        ] + tags
+        service_check_tags = list(set(service_check_tags))
 
         # Check for ssl configs and generate an appropriate ssl context object
         ssl_verify = instance.get('ssl_verify', True)
@@ -494,6 +496,7 @@ class VSphereCheck(AgentCheck):
         i_key = self._instance_key(instance)
         server_instance = self._get_server_instance(instance)
         perfManager = server_instance.content.perfManager
+        custom_tags = instance.get('tags', [])
 
         self.log.debug(
             "job_atomic: Querying available metrics"
@@ -517,7 +520,7 @@ class VSphereCheck(AgentCheck):
         self.morlist[i_key][mor_name]['last_seen'] = time.time()
 
         # ## <TEST-INSTRUMENTATION>
-        self.histogram('datadog.agent.vsphere.morlist_process_atomic.time', t.total())
+        self.histogram('datadog.agent.vsphere.morlist_process_atomic.time', t.total(), tags=custom_tags)
         # ## </TEST-INSTRUMENTATION>
 
     def _cache_morlist_process(self, instance):
@@ -662,7 +665,7 @@ class VSphereCheck(AgentCheck):
                 )
 
         # ## <TEST-INSTRUMENTATION>
-        self.histogram('datadog.agent.vsphere.metric_colection.time', t.total())
+        self.histogram('datadog.agent.vsphere.metric_colection.time', t.total(), tags=custom_tags)
         # ## </TEST-INSTRUMENTATION>
 
     def collect_metrics(self, instance):