+++ cassandra/CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG - cassandra
 
-1.2.1 / Unreleased
+1.2.1 / 2018-03-23
 ==================
 
 ### Changes