+++ hdfs_datanode/CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG - hdfs_datanode
 
-1.2.0 / Unreleased
+1.2.0 / 2018-03-23
 ==================
 
 ### Changes