+++ kyototycoon/CHANGELOG.md
@@ -1,5 +1,5 @@
 # CHANGELOG - kyototycoon
-1.2.0 / Unreleased
+1.2.0 / 2018-03-23
 ==================
 
 ### Changes