+++ lighttpd/CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG - lighttpd
 
-1.1.0 / Unreleased
+1.1.0 / 2018-03-23
 ==================
 
 ### Changes